Allow sandbox PID namespace to be used
Identify the sandbox PID namespace path and translate it to the root
PID namespace. This is one step towards supporting K8s ephemeral
container target. Currently, the target always comes as the sandbox
because containerd uses the process PID (in gVisor's case the sandbox
PID) to identify the container. Instead of landing in the desired
container, the root namespace is targeted to allow all processes to
be visible and debugged.

PiperOrigin-RevId: 531033866